requirements.txt
beautifulsoup4
lxml
Pillow
python-dotenv
selenium
webdriver-manager
requests
undetected-chromedriver
setuptools
